Fix provisioning of rubies
I'm maybe misreading the documentation but I'm not sure rvm::system should be required to install rubies for a given user. Without it error 'rvm install not found' causes a silent failure. I suspect the provisioner, as root, is potentially using the user's RVM binary to install a ruby and this is using internals from root's path instead of the users. Its only a hunch, I don't understand the cookbook internals enough and this works for now.